Add a not null check to avoid a NPE on LTD
In the loans to directors question page if a user
selects no to loans to directors, but has a ltd
additional information resource, it will attempt
to delete the loans resource but will fail with NPE
as it doesn't exist. Added a null check to avoid this.

+++ src/main/java/uk/gov/companieshouse/web/accounts/controller/smallfull/LoansToDirectorsQuestionController.java
@@ -96,8 +96,10 @@ public class LoansToDirectorsQuestionController extends BaseController {
             } else {
                 if (loansToDirectorsApi != null) {
                     if (loansToDirectorsApi.getLinks().getAdditionalInformation() != null) {
-                        for (String loanId : loansToDirectorsApi.getLoans().keySet()) {
-                            loansService.deleteLoan(transactionId, companyAccountsId, loanId);
+                        if (loansToDirectorsApi.getLoans() != null) {
+                            for (String loanId : loansToDirectorsApi.getLoans().keySet()) {
+                                loansService.deleteLoan(transactionId, companyAccountsId, loanId);
+                            }
                         }
                     } else {
                         loansToDirectorsService.deleteLoansToDirectors(transactionId, companyAccountsId);
